Refactoring in TTMetal dialect
This commit refactors:
1. Dialect conversion from TTKernel to EmitC.
2. Serialization of TTMetal IR to flatbuffer binary.

1. Implement dialect conversion from TTKernel to EmitC

TTKernel dialect that can be found nested in TTMetal ops can now be
converted via 'convert-ttkernel-to-emitc' pass. Pass is registered as a
func::FuncOp pass so the kernel must be put inside a function before
conversion.
When serializing ttmetal IR to binary, we call this conversion for every
region of a ttmetal dispatch op.

FileCheck UT is added.

2. Translate TTMetal to flatbuffer

Serialization to flatbuffer binary is now a proper translation pass that
can be run with:

ttmlir-translate --ttmetal-to-flatbuffer ttmetal.mlir
