+++ include/ttmlir/Dialect/TTMetal/Transforms/Passes.td
@@ -14,15 +14,4 @@ def ConvertTTIRToTTMetal: Pass<"convert-ttir-to-ttmetal", "::mlir::ModuleOp"> {
   }];
 }
 
-def TTMetalSerializeToBinary: Pass<"ttmetal-serialize-to-binary", "::mlir::ModuleOp"> {
-  let summary = "";
-  let description = [{
-    todo
-  }];
-
-  list<Option> options = [
-    Option<"output", "output", "std::string", "", "Output binary path">,
-  ];
-}
-
 #endif
