+++ lib/Dialect/TTMetal/Transforms/CMakeLists.txt
@@ -1,7 +1,5 @@
 add_mlir_dialect_library(MLIRTTMetalTransforms
         Passes.cpp
-        KernelsToCpp.cpp
-        SerializeToBinary.cpp
 
         ADDITIONAL_HEADER_DIRS
         ${PROJECT_SOURCE_DIR}/include/ttmlir
