+++ tools/ttmlir-translate/ttmlir-translate.cpp
@@ -5,6 +5,7 @@
 #include "mlir/InitAllTranslations.h"
 #include "mlir/Support/LogicalResult.h"
 #include "mlir/Tools/mlir-translate/MlirTranslateMain.h"
+#include <mlir/IR/DialectRegistry.h>
 
 using namespace mlir;
 
@@ -12,10 +13,15 @@ namespace mlir::tt::ttnn {
 void registerTTNNToFlatbuffer();
 } // namespace mlir::tt::ttnn
 
+namespace mlir::tt::ttmetal {
+void registerTTMetalToFlatbuffer();
+} // namespace mlir::tt::ttmetal
+
 // Place to register all the custom translations
 static void registerCustomTranslations() {
   static bool initOnce = []() {
     mlir::tt::ttnn::registerTTNNToFlatbuffer();
+    mlir::tt::ttmetal::registerTTMetalToFlatbuffer();
     return true;
   }();
   (void)initOnce;
@@ -24,5 +30,6 @@ static void registerCustomTranslations() {
 int main(int argc, char **argv) {
   registerAllTranslations();
   registerCustomTranslations();
+
   return failed(mlirTranslateMain(argc, argv, "MLIR Translation Testing Tool"));
 }
